Fixed linters by renaming
